requirements.txt
pandas
langchain-core
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-ollama
faiss-cpu
pypdf
sentence-transformers
python-dotenv
